Add ability to register/identify the nick
Add 'password <password>' to the server section of the configuration
No attempt to encrypt the password as it is passed somewhat plain text across
the pipe anyway

This does a very basic job looking for the server to say that the nick is
registered and that you need to identify yourself.  When it receives that
message, it sends the expected /msg NickServ identify <password> message.

This is to support FreeNode.