Retroactively op users when op priv received
When the bot receives op privs, it will check who is online in the channel that
doesn't have op privs which it would normally provide and does so.   THus, one
needs only to op the bot and everyone else will automatically get ops as
configured.

When two bots are in a channel (providing each other backup),  it attempts to
smartly remove the nick the other bot op'd so as not to repeat it.